+++ packages/adapter-prisma/src/index.ts
@@ -1,16 +1,22 @@
-import type { PrismaClient } from "@prisma/client";
-import type { Adapter, AdapterFunction } from "lucia-auth";
+import type { Adapter, AdapterFunction, SessionSchema, UserSchema } from "lucia-auth";
 import { getUpdateData } from "lucia-auth/adapter";
 import { convertSession } from "./utils.js";
+import { PrismaClient, SmartPrismaClient } from "./prisma.js";
 
 interface PossiblePrismaError {
 	code: string;
 	message: string;
 }
 
+type Schemas = {
+	user: UserSchema;
+	session: SessionSchema;
+};
+
 const adapter =
-	(prisma: PrismaClient): AdapterFunction<Adapter> =>
+	(prismaClient: PrismaClient<Schemas>): AdapterFunction<Adapter> =>
 	(LuciaError) => {
+		const prisma = prismaClient as any as SmartPrismaClient<Schemas>;
 		return {
 			getUser: async (userId) => {
 				const data = await prisma.user.findUnique({
@@ -119,7 +125,7 @@ const adapter =
 				}
 			},
 			deleteSession: async (sessionId) => {
-				await prisma.session.delete({
+				const t = await prisma.session.delete({
 					where: {
 						id: sessionId
 					}