+++ packages/adapter-prisma/src/utils.ts
@@ -1,7 +1,7 @@
-import { Session } from "@prisma/client";
 import type { SessionSchema } from "lucia-auth";
 
-export const convertSession = (session: Session): SessionSchema => {
+// TODO: Remove bigint => number conversion on next major version
+export const convertSession = (session: SessionSchema): SessionSchema => {
 	const { expires, idle_expires: idleExpires, ...data } = session;
 	return {
 		expires: Number(expires),