+++ packages/adapter-prisma/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 0.2.2
+
+- [Fix] Remove minor breaking change introduced in `0.2.1`
+
+## 0.2.1
+
+- Remove `prisma` and `@prisma/client` dependencies
+
 ## 0.2.0
 
 - [Breaking] Require `lucia-auth` 0.4.0